Moved render methods to base clas
